requirements.txt
CherryPy==18.6.1
cherrypy_cors==1.6
numpy==1.17.4
Pillow==9.0.0
torch==1.10.1
torchvision==0.11.2
